requirements.txt
flask==3.0.0
yt-dlp==2024.3.10
instaloader==4.10
megapy==1.0.0
gdown==5.1.0
requests==2.31.0
internetarchive==3.5.0
requests-toolbelt==1.0.0
pycryptodome==3.19.0
ffmpeg-python==0.2.0
schedule==1.2.0
curl-cffi==0.6.2
selenium==4.16.0
webdriver-manager==4.0.1
telethon==1.34.0
Pillow==10.2.0
gunicorn==21.2.0
